/ taskget: use has/hasnt instead of word/noword to test for tags
the '+tag' search filter was not working because tasklib adds this as
"tags.word:'word'" not "tags.word:word", ie, it adds a single quote.
this works with 'has' but not with 'word'

this is actually an imperfect solution because it's a substring match.
we should probably use a better technique but this is at least more
functional for now