/ taskget: really fix +/- tag filter, needs to be whole tag match
originally we used tags__word and tags__noword, which ended up as
"tags.word:'word'" due to tasklib.  the quotes made it work wrong.
since we will eventually move away from tasklib and use taskchampion
instead for direct db access, and because tasklib project seems to be
essentially abandoned, there's no sense fixing this in tasklib.

so as the next attempt, we used tags__has and tags__hasnt, in the last
commit, which works with the quotes (don't know why it's not affected by
the same quote issue as tags.word) but actually tests for substrings
rather than the whole "word" so like a "+pre" would end up matching both
"+preset" and "+preliminary"

tasklib however does accept strings in addition to dicts as arguments
for the filter, which it will provide unchanged on the command line.  so
we can just give strings like "+preset" as additional args to .filter()
method and it then it works correctly