/ taskget: correctly handle [short] uuids that are all base10 digits
we were using int(foo) ValueError to see if it was an int, if so we
assumed it's a taskid, do a lookup and if no match, no such uuid.

actually large integer could be a uuid prefix that happens to have all
base10 digits.  such is the case with our task with [short] uuid
24601392.

what we do now is, if it's an integer, but doesn't look up, and also
it's 8 characters or more, we assume it's a uuid and allow to continue
rather than failing with no match.  this is technically imperfect, but
in practice will be fine: even the lowest 8 digit number is 10 million
and if someone has that many tasks, they have quite a few other problems
before hitting this one ;-)