+ task: add inv wrapper that always sets rc.defaultwidth=$COLUMNS
this helps when we use pipes.  and can still override on command line
with rc.defaultwidth=0 when we want for a more parsable export view

see taskwarrior issue 991 to see why we need this

we borrowed the invocation paradigm -- removing our own path from path
and re-lookup -- from our tmux wrapper, we do the same thing there.
probably we should generalize that, TODO