+ taskinfo: new task.sh invocation that wraps and condenses "task info"
started just wanting to remove the description -- which ended up with
very many lines when large note counts -- but ended up making this to
just fit more with the ideal.  maybe overkill and certainly a hack
because we're scraping curses output.  but for this moment this should
be a fine hack.  eventually we should probably access the db directly
but that's likely a long way off for us

ideally taskwarrior would have some mode or flag that would only show
the first line of description rather than all annotations