/ tasksh: fix "task $uuid" needs "uuid:$uuid" else sometimes string math
for some reason uuids like 73610118-7e66-4ba0-a8cf-c26f66e6f5b2 make
taskwarrior act like we wanted it to do string math and the '-' are
minus.  this is probably a bug, but we'll just have to start treating
uuid truly like a string field with a name uuid everywhere and that
should fix it for now.

we'll consider if we think this is a bug that the uuid detection
heuristic is broken.  probably need to test on later version of
taskwarrior before concluding anything