+ task2to3: import a tw3 db via tw2 undo.data, preserving change history
at first the changelog shown by rc.journal.info=1 in v2 was not
implemented in the 3.0 release, a significant regression.  3.3 added
support for it, in a slightly different form (initial state is shown,
and tag changes are displayed individually).

the problem, however, is that, on import of a full task export, all
intermediate changes are lost forever: only the end state of the task is
imported, and the tasks's lifetime last-update timestamp.  even if the
individual changes weren't lost by import/export, they would not have
correct timestamps, because taskwarrior simply uses the current time to
generate it.  for a big import (like v2 to v3 upgrade), the history
would show all the changes as happening at the same time, and would not
show intermediate states.  this results in a considerable loss of
information when upgrading from v2 to v3.

instead, here is a new script (task2to3.py) that can reconstitute the
taskwarrior database by replaying the undo log for importation, forcing
each recorded event to happen at the logged [in undo.data] time by
importing one change at a time, and using "faketime" to execute
the "taskwarrior import".

this is still quite a hack, because the change history is not actually a
part of the import/export interface; it seems to be generated internally
by examining the system time, with no way to override or hint (besides
faking the time, as we do here).  in the future, a more general fix
would be better.  probably we could allow a changetime override, either
within the json record being imported, or at least as a cli arg to
"import" (latter forces one invocation per record though).  however, the
versions and timestamps would still not be "export" data.  not sure how
to address this without adding a history table to sqlite, but that
violates the backend abstraction, and keeping the records within the
json export would result in a dramatic increase in size, and make manual
record inspection at the shell (ie, "task export | jq") unwieldy.

we'll punt on that one for now...