/ taskget: for bare integers, keep searching if id lookup fails
we had previously assumed that a naked integer was necessarily an id
number and if it didn't match, would simply move on to the next
argument.  this doesn't work when we are looking for years or other
substrings that would be in labels or descriptions, etc.  for example we
use label ota202505 for applying may phone update, but search for 202505
will fail and there would be no results.

this will make a lot more matches in certain cases, probably there
should be a flag to control the behavior, but we'll try it like this for
the moment and see how it goes