/ taskdeps: work around taskwarrior dependencies being compared as strings
taskwarrior is using startswith on the comma-separated list of
dependencies for matching, which is pretty bogus.  we can sort of hack
around it with .has to at least check the uuid as a substring.  see the
upstream issues and our own tasks from accompanying comment in the code
for more details