- readme: tasklib: add note about future plans to kill tasklib reliance
tasklib is dead, so we should stop using it.  we make minimal use of it
anyways and this would be easy enough to slurp and cache.  all it does
is parse a 'task export' and give it ORM-like semantics which we don't
actually care about.  We could easily do this ourselves.

Another option is to slurp it into duckdb using its read_json() like
we're doing in cmdhist.  That turns out to be pretty fast and useful,
gives a lot of options, and duckdb is well supported and active project,
unlike tasklib.