+ cmdhist: add script we use to join bash cli history with timew data
this turns out to be pretty useful because command line history is just
a big heap of untagged stuff and this gets us nice per-task and
per-project visibility that can be whittled by date.  it's very early
implementation at the moment but has already proven highly useful on
several occasions.